Lexer: Resolve string literal error message issues
Prior to this commit the scanning of strings would be aborted mid-scan
if an error was found. This could later result in an uneven amount of
double-quotes in a program, as well as incorrect tokens being output.

By continuing to scan a string even after an error is found this issue
was eliminated.